Moved ajax tracking to postlist module and added backend validation for row id
